VNNewsVoice_MLWorkerService/pyproject.toml
[project]
name = "vnnewsvoice-mlworker-service"
version = "0.1.0"
description = "Queue-only worker for summarization and TTS in VNNewsVoice pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika==9.4.3",
    "boto3==1.40.18",
    "cloudinary==1.44.1",
    "fastapi==0.115.14",
    "google-genai==1.26.0",
    "huggingface-hub>=0.36.2",
    "mlflow>=2.10.0",
    "peft>=0.7.1",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "torch==2.7.1+cpu",
    "transformers==4.53.1",
    "uvicorn[standard]==0.35.0",
    "prometheus-fastapi-instrumentator>=0.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "pytest>=8.4.2",
]
mlops = [
    "datasets>=4.1.0",
    "dvc[s3]>=3.63.0",
    "rouge-score>=0.1.2",
    "bitsandbytes>=0.47.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]

[tool.ruff]
exclude = ["models/", "*.ipynb"]
